// src/app/models/types/Quotations/additionalCovers.ts
export type AdditionalCoversType = {
    Electrical?: { checked?: boolean; validation?: string; value?: number }
    NonElectrical?: { checked?: boolean; validation?: string; value?: number }
    BiFuel?: { checked?: boolean; validation?: string; value?: number }
    AntiTheft?: { checked?: boolean; validation?: string; value?: number }
    AA?: { checked?: boolean; validation?: string; value?: number }
    VoluntaryExcess?: { checked?: boolean; validation?: string; value?: number }
}